requirements.txt
pillow 
moviepy